[wasm][debugger][tests] Extract EvaluateOnCallFrame tests to new file
